rename wmdocker -> docker
