+++ internal/engine/build_and_deployer_test.go
@@ -18,13 +18,13 @@ import (
 	"github.com/docker/docker/api/types"
 	"github.com/stretchr/testify/assert"
 	"github.com/windmilleng/tilt/internal/build"
+	"github.com/windmilleng/tilt/internal/docker"
 	"github.com/windmilleng/tilt/internal/k8s"
 	"github.com/windmilleng/tilt/internal/model"
 	"github.com/windmilleng/tilt/internal/synclet"
 	"github.com/windmilleng/tilt/internal/testutils"
 	"github.com/windmilleng/tilt/internal/testutils/output"
 	"github.com/windmilleng/tilt/internal/testutils/tempdir"
-	"github.com/windmilleng/tilt/internal/wmdocker"
 	"github.com/windmilleng/wmclient/pkg/dirs"
 )
 
@@ -134,7 +134,7 @@ func TestIncrementalBuildFailure(t *testing.T) {
 
 	ctx := output.CtxForTest()
 
-	f.docker.ExecErrorToThrow = wmdocker.ExitError{ExitCode: 1}
+	f.docker.ExecErrorToThrow = docker.ExitError{ExitCode: 1}
 	_, err := f.bd.BuildAndDeploy(ctx, SanchoManifest, NewBuildState(alreadyBuilt))
 	msg := "Command failed with exit code: 1"
 	if err == nil || !strings.Contains(err.Error(), msg) {
@@ -340,7 +340,7 @@ RUN ["go", "install", "github.com/windmilleng/sancho"]`,
 type bdFixture struct {
 	*tempdir.TempDirFixture
 	ctx    context.Context
-	docker *wmdocker.FakeDockerClient
+	docker *docker.FakeDockerClient
 	k8s    *FakeK8sClient
 	bd     BuildAndDeployer
 }
@@ -363,7 +363,7 @@ func newBDFixture(t *testing.T, env k8s.Env) *bdFixture {
 func newBDFixtureHelper(t *testing.T, env k8s.Env, fallbackFn FallbackTester) *bdFixture {
 	f := tempdir.NewTempDirFixture(t)
 	dir := dirs.NewWindmillDirAt(f.Path())
-	docker := wmdocker.NewFakeDockerClient()
+	docker := docker.NewFakeDockerClient()
 	docker.ContainerListOutput = map[string][]types.Container{
 		"pod": []types.Container{
 			types.Container{
