+++ internal/synclet/wire_gen.go
@@ -7,8 +7,9 @@ package synclet
 
 import (
 	context "context"
+
+	wmdocker "github.com/windmilleng/tilt/internal/docker"
 	k8s "github.com/windmilleng/tilt/internal/k8s"
-	wmdocker "github.com/windmilleng/tilt/internal/wmdocker"
 )
 
 // Injectors from wire.go:
